additional refactor hack
